pyproject.toml
[project]
name = "pretrain-vlm"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0,<2",
    "beautifulsoup4>=4.12.0",
    "datasets>=3.6.0,<5",
    "gdown>=6.0.0,<7",
    "jinja2>=3.1,<4",
    "loguru>=0.7.3",
    "openai>=2.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=12.2.0,<13",
    "pyyaml>=6.0.3,<7",
    "requests>=2.32.0",
    "torch>=2.10.0,<3",
    "transformers>=4.55.4,<5",
]

[tool.ruff]
target-version = "py311"
line-length = 110

[tool.ruff.format]
skip-magic-trailing-comma = true
